Simplify rubysl/complex
