--- lib/rubysl/rubysl-complex/lib/complex.rb
@@ -1,1 +1,0 @@
-require "rubysl/complex"
