--- lib/rubysl/rubysl-complex/lib/rubysl/complex.rb
@@ -1,2 +1,0 @@
-require "rubysl/complex/complex"
-require "rubysl/complex/version"
