--- lib/rubysl/rubysl-complex/lib/rubysl/complex/complex.rb
@@ -1,26 +1,0 @@
-# :enddoc:
-
-require 'cmath'
-
-unless defined?(Math.exp!)
-  Object.instance_eval{remove_const :Math}
-  Math = CMath
-end
-
-def Complex.generic? (other)
-  other.kind_of?(Integer) ||
-  other.kind_of?(Float)   ||
-  other.kind_of?(Rational)
-end
-
-class Complex
-
-  alias image imag
-
-end
-
-class Numeric
-
-  def im() Complex(0, self) end
-
-end
