--- lib/rubysl/rubysl-complex/lib/rubysl/complex/version.rb
@@ -1,5 +1,0 @@
-module RubySL
-  module Complex
-    VERSION = "2.0.0"
-  end
-end
